server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "graypy>=2.1.0",
    "jinja2>=3.1.6",
    "boto3>=1.37.18",
    "kafka-python>=2.2.6",
    "prometheus-client>=0.21.1",
    "pyyaml>=6.0.2",
    "uvicorn>=0.34.0",
    "pydantic[email,timezone]>=2.11.0",
    "sqlalchemy>=2.0.39",
    "psycopg2-binary>=2.9.10",
    "yoyo-migrations>=9.0.0",
    "asyncpg>=0.30.0",
    "python-multipart>=0.0.20",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib == 1.7.4",
    "bcrypt == 4.1.3",
    "pillow (>=11.2.1,<12.0.0)",
    "reportlab (>=3.6) ; python_version < \"4.0\"",
    "httpx>=0.28.1",
]

[dependency-groups]
dev = [
    "beautifulsoup4>=4.13.3",
    "lxml>=5.3.1",
    "pyright>=1.1.397",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "pytest-cov>=6.0.0",
    "pytest-html>=4.1.1",
    "ruff>=0.11.0",
]

[tool.ruff]
line-length = 160
target-version = "py313"
exclude = [
]


[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.pytest.ini_options]
pythonpath = ["./src", "./tests"]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
log_cli = true
log_cli_level = "DEBUG"
addopts = "--cov=./src --cov-report=term --cov-report=html --html=pytest-report/index.html"


[tool.pyright]
venvPath = "."
venv = ".venv"
typeCheckingMode = "strict"
exclude = [
]
